Respond to Tianqi's Comments
Move IndexedGraph to it's own header, edit python imports
